CpuStats changes
Move CpuUsage into inner class
Add ThrottlingData as inner class
